requirements.txt
requests>=2.28.0
Pillow>=9.0.0
RPi.GPIO>=0.7.1
spidev>=3.5
